Rewrite RemittanceBridge with pull-based USDC flow and mock mode local transfer (64 tests passing)

+++ packages/contracts/contracts/mocks/SimpleMockBridge.sol
@@ -9,13 +9,13 @@ pragma solidity ^0.8.27;
 contract SimpleMockBridge {
     address public lastRecipient;
     uint256 public lastAmount;
-    uint32  public lastDestChainId;
+    uint256 public lastDestChainId;
     uint256 public callCount;
 
     function sendRemittance(
         address recipient,
         uint256 usdcAmount,
-        uint32  destChainId
+        uint256 destChainId
     ) external {
         lastRecipient   = recipient;
         lastAmount      = usdcAmount;
